requirements.txt
asgiref==3.9.1
Django==4.2.23
djangorestframework==3.16.0
joblib==1.5.1
numpy==2.0.2
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
